Refactor Codemaat Tests and code cleanup
