+++ analysis/import/CodeMaatImporter/build.gradle
@@ -1,34 +1,18 @@
-repositories {
-    maven {
-        url 'https://dl.bintray.com/spekframework/spek-dev'
-    }
-}
-
 dependencies {
     implementation project(':model')
     implementation project(':tools:InteractiveParser')
 
     implementation group: 'com.univocity', name: 'univocity-parsers', version: univocity_parsers_version
     implementation group: 'info.picocli', name: 'picocli', version: picocli_version
-    implementation group: 'io.fastjson', name: 'boon', version: boon_version
     implementation group: 'io.github.microutils', name: 'kotlin-logging', version: kotlin_logging_version
     implementation group: 'com.github.kotlin-inquirer', name: 'kotlin-inquirer', version: kotlin_inquirer_version
-
-    implementation group: 'org.slf4j', name: 'slf4j-simple', version: slf4j_version
 
     testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter', version: junit5_version
     testImplementation group: 'org.jetbrains.kotlin', name: 'kotlin-test', version: kotlin_version
     testImplementation group: 'org.assertj', name: 'assertj-core', version: assertj_version
-    testImplementation group: 'org.hamcrest', name: 'hamcrest-library', version: hamcrest_version
     testImplementation group: 'io.mockk', name: 'mockk', version: mockk_version
-    testImplementation("org.spekframework.spek2:spek-dsl-jvm:$spek2_version") {
-        exclude group: 'org.jetbrains.kotlin'
-    }
 
     testRuntimeOnly group: 'org.jetbrains.kotlin', name: 'kotlin-reflect', version: kotlin_version
-    testRuntimeOnly("org.spekframework.spek2:spek-runner-junit5:$spek2_version") {
-        exclude group: 'org.jetbrains.kotlin'
-    }
 }
 
 test {
